+++ src/util/custom-url-validator.ts
@@ -1,34 +1,39 @@
-import { Injectable } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
 import {
   isURL,
   ValidationArguments,
   ValidatorConstraint,
   ValidatorConstraintInterface,
 } from 'class-validator';
-import { GiftogetherExceptions } from 'src/filters/giftogether-exception';
+import * as ValidatorJs from 'validator';
+
+export function isUrlOptions(): ValidatorJs.IsURLOptions {
+  if (process.env['DEBUG'] === 'true') {
+    return {};
+  }
+  return {
+    require_protocol: true,
+    protocols: ['https'],
+    require_host: true,
+    require_port: false,
+    host_blacklist: [/\*/],
+    host_whitelist: [process.env['AWS_S3_HOST']],
+  };
+}
 
 /**
  * Custom validator implementation to check if a given string is a valid URL.
  * Uses the `isURL` function from the 'Validator' library with configuration options retrieved from the ConfigService.
  */
 @ValidatorConstraint({ name: 'CustomUrlValidator', async: true })
-@Injectable()
 export class CustomUrlValidator implements ValidatorConstraintInterface {
-  constructor(
-    private readonly g2gException: GiftogetherExceptions,
-    private readonly configService: ConfigService,
-  ) {}
-
-  validate(
+  async validate(
     url: string,
     _validationArguments?: ValidationArguments,
-  ): Promise<boolean> | boolean {
-    const options = this.configService.get('isUrlOptions');
-    return isURL(url, options);
+  ): Promise<boolean> {
+    return isURL(url, isUrlOptions());
   }
 
   defaultMessage?(_validationArguments?: ValidationArguments): string {
-    return this.g2gException.InvalidUrl.message;
+    return '유효한 URL이 아닙니다.';
   }
 }
